use ints to ensure whole numbers

+++ src/main/java/me/robomwm/MountainDewritoes/BetterZeldaHearts.java
@@ -202,8 +202,8 @@ public class BetterZeldaHearts implements Listener
             event.getPlayer().getAttribute(Attribute.GENERIC_MAX_HEALTH).setBaseValue(60D);
         else
         {
-            double extraHearts = event.getPlayer().getAttribute(Attribute.GENERIC_MAX_HEALTH).getBaseValue() - 60D;
-            event.getPlayer().getAttribute(Attribute.GENERIC_MAX_HEALTH).setBaseValue(60D + (extraHearts - (extraHearts/8)));
+            int extraHearts = (int)event.getPlayer().getAttribute(Attribute.GENERIC_MAX_HEALTH).getBaseValue() - 60;
+            event.getPlayer().getAttribute(Attribute.GENERIC_MAX_HEALTH).setBaseValue(60 + (extraHearts - (extraHearts/8)));
         }
     }
 }
